- rename and moved files to the correct folder
